Flytter testkode fra src/main/ til src/test
