+++ src/main/java/no/nav/tag/tiltaksgjennomforing/featuretoggles/FeatureToggleConfig.java
@@ -1,15 +1,13 @@
 package no.nav.tag.tiltaksgjennomforing.featuretoggles;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
 import no.finn.unleash.DefaultUnleash;
-import no.finn.unleash.FakeUnleash;
 import no.finn.unleash.Unleash;
 import no.finn.unleash.util.UnleashConfig;
 import no.nav.tag.tiltaksgjennomforing.featuretoggles.enhet.ByEnhetStrategy;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
 
 @Configuration
@@ -18,11 +16,11 @@ public class FeatureToggleConfig {
     private static final String APP_NAME = "tiltaksgjennomforing-api";
 
     @Bean
-    @Profile(value= {"prod", "preprod"})
+    @ConditionalOnProperty(name = "${tiltaksgjennomforing.unleash.unleash-uri}")
     public Unleash initializeUnleash(@Value(
-            "${tiltaksgjennomforing.unleash.unleash-uri}") String unleashUrl, 
-            ByEnvironmentStrategy byEnvironmentStrategy,
-            ByEnhetStrategy byEnhetStrategy) {
+            "${tiltaksgjennomforing.unleash.unleash-uri}") String unleashUrl,
+                                     ByEnvironmentStrategy byEnvironmentStrategy,
+                                     ByEnhetStrategy byEnhetStrategy) {
         UnleashConfig config = UnleashConfig.builder()
                 .appName(APP_NAME)
                 .instanceId(APP_NAME + "-" + byEnvironmentStrategy.getEnvironment())
@@ -34,13 +32,5 @@ public class FeatureToggleConfig {
                 byEnvironmentStrategy,
                 byEnhetStrategy
         );
-    }
-    
-    @Bean
-    @Profile(value= {"dev", "heroku"})
-    public Unleash unleashMock() {
-        FakeUnleash fakeUnleash = new FakeUnleash();
-        fakeUnleash.enableAll(); //Enabler alle toggles pr. default. Kan endres lokalt ved behov.
-        return fakeUnleash;
     }
 }
