+++ src/main/java/no/nav/tag/tiltaksgjennomforing/infrastruktur/SjekkAktiveProfilerInitializer.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class SjekkAktiveProfilerInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
 
-    public static final List<String> MILJOER = Arrays.asList("dev", "heroku", "preprod", "prod");
+    public static final List<String> MILJOER = Arrays.asList("preprod", "prod");
 
     @Override
     public void initialize(ConfigurableApplicationContext applicationContext) {
