--- src/main/resources/config/application-heroku.yaml
@@ -1,31 +1,0 @@
-spring:
-  datasource:
-    platform: postgres
-    driverClassName: org.postgresql.Driver
-
-no.nav.security.oidc:
-  issuers: selvbetjening, isso
-  issuer:
-    selvbetjening:
-      discoveryurl: http://metadata-selvbetjening
-      accepted_audience: aud-selvbetjening
-    isso:
-      discoveryurl: http://metadata-isso
-      accepted_audience: aud-isso
-
-tiltaksgjennomforing:
-  altinn-tilgangsstyring:
-    uri: http://localhost:8090
-    altinn-api-key: test
-    api-gw-api-key: test
-    service-code: 42
-    service-edition: 1
-  ereg:
-    uri: http://localhost:8090
-  sts:
-    username: foo
-    password: bar
-    ws-uri: http://foo.bar
-    rest-uri: http://localhost:8090/sts/sts/token
-  abac:
-    veilarbabac-uri: http://localhost:8090/veilarbabac
