+++ src/test/java/no/nav/tag/tiltaksgjennomforing/IntegrasjonerMockServer.java
@@ -1,4 +1,4 @@
-package no.nav.tag.tiltaksgjennomforing.infrastruktur;
+package no.nav.tag.tiltaksgjennomforing;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
