+++ src/test/resources/logback-spring.xml
@@ -1,10 +1,11 @@
 <configuration>
+    <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
     <appender name="consoleAppender" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder class="net.logstash.logback.encoder.LogstashEncoder"/>
+        <encoder>
+            <pattern>${CONSOLE_LOG_PATTERN}</pattern>
+            <charset>utf8</charset>
+        </encoder>
     </appender>
-    <logger name="jsonLogger" additivity="false" level="DEBUG">
-        <appender-ref ref="consoleAppender"/>
-    </logger>
     <logger name="no.nav.vault.jdbc.hikaricp.VaultUtil" level="WARN"/>
     <root level="INFO">
         <appender-ref ref="consoleAppender"/>
