+++ src/test/resources/mappings/altinn_varsel.json
