+++ src/test/resources/mappings/axsys.json
