+++ src/test/resources/mappings/ereg.json
