+++ src/test/resources/mappings/sts.json
