+++ src/test/resources/metadata-selvbetjening.json
